requirements.txt
asgiref==3.2.5
astroid==2.3.3
colorama==0.4.3
Django==3.0.4
django-appconf==1.0.3
django-bootstrap-modal-forms==1.5.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-select2==7.2.3
Faker==4.0.3
freeze==1.0.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
typed-ast==1.4.1
wrapt==1.11.2
